refactor(ui): refactor Notification and add stories
