--- ui/design/src/components/Notification/index.ts
@@ -1,5 +1,0 @@
-import { Notification, notify } from './toasts/notification-toast';
-import BookmarkPill from './notification-pills/bookmark-pill';
-import NewPostsPill from './notification-pills/new-posts-pill';
-
-export { Notification, notify, BookmarkPill, NewPostsPill };
