+++ ui/design/src/components/StackedAvatar/StackedAvatar.stories.tsx
@@ -4,30 +4,12 @@ import { Grommet } from 'grommet';
 import StackedAvatar from '.';
 
 import lightTheme from '../../styles/themes/light/light-theme';
+import { userData } from '../../utils/dummy-data';
 
 export default {
   title: 'Avatars/StackedAvatar',
   component: StackedAvatar,
 };
-
-const userData: { ethAddress: string; avatar?: string }[] = [
-  {
-    ethAddress: '0x003410490050000320006570034567114572000',
-    avatar: 'https://placebeard.it/360x360',
-  },
-  {
-    ethAddress: '0x004410490050000320006570034567114572001',
-    avatar: 'https://placebeard.it/360x360',
-  },
-  {
-    ethAddress: '0x005410490050000320006570034567114572002',
-    avatar: 'https://placebeard.it/360x360',
-  },
-  {
-    ethAddress: '0x006410490050000320006570034567114572003',
-    avatar: 'https://placebeard.it/360x360',
-  },
-];
 
 const Template = (args: any) => (
   <Grommet theme={lightTheme}>
